Move all tasks to Gruntfile.js and eliminate grunt/tasks/build.js. Trim package.json.
